+++ src/components/UI/form-items/inline-input.component.js
@@ -3,7 +3,7 @@ import { OutlinedInput, InputLabel, FormControl, FormHelperText } from "@materia
 
 import { Button } from "../button.component";
 
-import classes from "../modules/form/inline-input.module.scss";
+import classes from "../modules/form-items/inline-input.module.scss";
 
 export const InlineInput = ({ type, name, value, disabled, error, touched, buttonLabel, className, buttonType, onClick, helperText, label, ...rest }) => {
   return (
