+++ src/components/UI/form-items/phone-input.component.js
@@ -3,7 +3,7 @@ import { FormControl, FormHelperText } from "@material-ui/core";
 import PhoneInput from "react-phone-input-2";
 
 import "react-phone-input-2/lib/material.css";
-import classes from "../modules/form/phone-input.module.scss";
+import classes from "../modules/form-items/phone-input.module.scss";
 
 export const InputPhone = ({ onChange, value, country, error }) => {
   return (
