+++ src/components/UI/modules/form-items/switch.module.scss
